requirements.txt
Django~=5.1.6
djangorestframework~=3.15.2
celery~=5.4.0
python-telegram-bot~=21.10
django-filter~=25.1
drf-spectacular~=0.28.0
pillow~=11.1.0
stripe~=11.6.0
python-dotenv~=1.0.1
pandas~=2.2.3
asgiref~=3.8.1